requirements.txt
fastapi==0.111.0
google_api_python_client==2.136.0
google_auth_oauthlib==1.2.0
msal==1.29.0
O365==2.0.36
protobuf==5.27.2
python-dotenv==1.0.1
Requests==2.32.3
